backend/requirements.txt
asgiref==3.7.2
Django==4.1
django-cors-headers==4.2.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
mysqlclient==2.2.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
whitenoise==6.5.0
